requirements.txt
sphinx>=8.1.3
sphinx_rtd_theme>=3.0.2
myst_parser>=4.0.0
graphdatascience
retry
streamlit
numpy
pandas
plotly
py2neo

aiohappyeyeballs==2.4.6
 aiohttp==3.11.12
 aiosignal==1.3.2
 altair==5.5.0
 annotated-types==0.7.0
 anyio==4.8.0
 async-timeout==4.0.3
 attrs==25.1.0
 blinker==1.9.0
 cachetools==5.5.2
 certifi==2025.1.31
 charset-normalizer==3.4.1
 click==8.1.8
 dataclasses-json==0.6.7
 distro==1.9.0
 exceptiongroup==1.2.2
 frozenlist==1.5.0
 gitdb==4.0.12
 GitPython==3.1.44
 h11==0.14.0
 httpcore==1.0.7
 httpx==0.28.1
 httpx-sse==0.4.0
 idna==3.10
 iniconfig==2.0.0
 Jinja2==3.1.6
 jiter==0.8.2
 jsonpatch==1.33
 jsonpointer==3.0.0
 jsonschema==4.23.0
 jsonschema-specifications==2024.10.1
 langchain==0.3.19
 langchain-community==0.3.17
 langchain-core==0.3.35
 langgraph==0.2.73
 langgraph-checkpoint==2.0.15
 langgraph-sdk==0.1.51
 langsmith==0.3.8
 MarkupSafe==3.0.2
 marshmallow==3.26.1
 msgpack==1.1.0
 multidict==6.1.0
 mypy-extensions==1.0.0
 narwhals==1.32.0
 neo4j==5.28.1
 numpy==1.26.4
 openai==1.63.0
 orjson==3.10.15
 packaging==24.2
 pandas==2.2.3
 pillow==11.1.0
 pluggy==1.5.0
 propcache==0.2.1
 protobuf==5.29.4
 pyarrow==19.0.1
 pydantic==2.10.6
 pydantic-settings==2.7.1
 pydantic_core==2.27.2
 pydeck==0.9.1
 pytest==8.3.4
 python-dateutil==2.9.0.post0
 python-dotenv==1.0.1
 pytz==2025.2
 PyYAML==6.0.2
 referencing==0.36.2
 regex==2024.11.6
 requests==2.32.3
 requests-toolbelt==1.0.0
 rpds-py==0.23.1
 six==1.17.0
 smmap==5.0.2
 sniffio==1.3.1
 SQLAlchemy==2.0.38
 streamlit==1.43.2
 tenacity==9.0.0
 tiktoken==0.9.0
 toml==0.10.2
 tomli==2.2.1
 tornado==6.4.2
 tqdm==4.67.1
 typing-inspect==0.9.0
 typing_extensions==4.12.2
 tzdata==2025.2
 urllib3==2.3.0
 yarl==1.18.3
 zstandard==0.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
beautifulsoup4==4.13.3
boto3==1.37.9
botocore==1.37.9
capstone==0.1
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.6
jmespath==1.0.1
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.20
langchain-aws==0.2.15
langchain-community==0.3.19
langchain-core==0.3.43
langchain-experimental==0.3.4
langchain-neo4j==0.4.0
langchain-text-splitters==0.3.6
langgraph==0.3.5
langgraph-checkpoint==2.0.18
langgraph-prebuilt==0.1.2
langgraph-sdk==0.1.55
langsmith==0.3.13
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
mypy-extensions==1.0.0
neo4j==5.28.1
neo4j-graphrag==1.6.0
networkx==3.4.2
newsapi-python==0.2.7
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
propcache==0.3.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyparsing==3.2.1
pypdf==5.4.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.11.4
setuptools==76.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sympy==1.13.1
tenacity==9.0.0
torch==2.6.0
triton==3.2.0
types-PyYAML==6.0.12.20241230
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wikipedia==1.4.0
yarl==1.18.3
zstandard==0.23.0

